requirements.txt
# Flask API Server
Flask>=2.3.0
Flask-CORS>=4.0.0
werkzeug>=2.3.0
gunicorn>=21.0.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# Utilities
tqdm>=4.66.0
requests>=2.31.0 